chore: starter/nuxt + examples/nuxt-2-composition-api
