+++ examples/nuxt-2-composition-api/.env.example
@@ -1,0 +1,10 @@
+# Nacelle
+NACELLE_SPACE_ID=<your-nacelle-space-id>
+NACELLE_GRAPHQL_TOKEN=<your-nacelle-public-token>
+NACELLE_ENDPOINT=<your-nacelle-storefront-graphql-endpoint>
+NACELLE_LOCALE=<active-locale>
+
+# Shopify Checkout (for details, see: https://www.npmjs.com/package/@nacelle/shopify-checkout)
+MYSHOPIFY_DOMAIN=<your-shop-id>
+SHOPIFY_STOREFRONT_CHECKOUT_TOKEN=<your-storefront-api-token>
+SHOPIFY_STOREFRONT_API_VERSION=<storefront-api-version-of-choice>
