+++ examples/nuxt-2-composition-api/.eslintrc.js
@@ -11,7 +11,5 @@ module.exports = {
   extends: ['@nuxtjs', 'plugin:nuxt/recommended', 'prettier'],
   plugins: [],
   // add your custom rules here
-  rules: {
-    'no-console': 'off'
-  }
+  rules: {}
 };
