+++ examples/nuxt-2-composition-api/.gitignore
