+++ examples/nuxt-2-composition-api/components/Collection.vue
