+++ examples/nuxt-2-composition-api/components/Product.vue
