+++ examples/nuxt-2-composition-api/components/Search.vue
