+++ examples/nuxt-2-composition-api/layouts/default.vue
