+++ examples/nuxt-2-composition-api/pages/404.vue
@@ -1,0 +1,3 @@
+<template>
+  <div>404</div>
+</template>
