+++ examples/nuxt-2-composition-api/pages/index.vue
@@ -14,12 +14,6 @@
   </div>
 </template>
 
-<script>
-export default {
-  name: 'HomePage'
-};
-</script>
-
 <style lang="scss" scoped>
 .home {
   align-items: center;
