+++ examples/nuxt-2-composition-api/pages/search.vue
