+++ examples/nuxt-2-composition-api/plugins/globalSetup.js
