+++ examples/nuxt-2-composition-api/store/README.md
