--- starters/nuxt-2/.env.example
@@ -1,9 +1,0 @@
-# Nacelle
-NACELLE_STOREFRONT_ENDPOINT="<YOUR_NACELLE_STOREFRONT_ENDPOINT>"
-NACELLE_STOREFRONT_TOKEN="<YOUR_NACELLE_PUBLIC_STOREFRONT_TOKEN>"
-NACELLE_STOREFRONT_LOCALE="<YOUR_NACELLE_STOREFRONT_LOCALE>"
-
-# Shopify Checkout (for details, see: https://www.npmjs.com/package/@nacelle/shopify-checkout)
-MYSHOPIFY_DOMAIN="<YOUR_MYSHOPIFY_DOMAIN>"
-SHOPIFY_STOREFRONT_CHECKOUT_TOKEN="<YOUR_SHOPIFY_STOREFRONT_CHECKOUT_TOKEN>"
-SHOPIFY_STOREFRONT_API_VERSION="<YOUR_SHOPIFY_STOREFRONT_API_VERSION>"
