--- starters/nuxt-2/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "nuxt-2",
-  "version": "1.0.0",
-  "private": true,
-  "scripts": {
-    "dev": "nuxt",
-    "build": "nuxt build",
-    "start": "nuxt start",
-    "generate": "nuxt generate",
-    "lint:js": "eslint --ext \".js,.vue\" --ignore-path .gitignore .",
-    "lint": "npm run lint:js",
-    "lintfix": "npm run lint:js -- --fix"
-  },
-  "dependencies": {
-    "@nacelle/shopify-checkout": "^0.0.8",
-    "@nacelle/storefront-sdk": "^1.0.3",
-    "core-js": "^3.19.3",
-    "fuse.js": "^6.5.3",
-    "idb-keyval": "^6.1.0",
-    "nuxt": "^2.15.8",
-    "uuid": "^8.3.2",
-    "vue": "^2.6.14"
-  },
-  "devDependencies": {
-    "@babel/eslint-parser": "^7.16.3",
-    "@nuxt/image": "^0.6.0",
-    "@nuxtjs/eslint-config": "^8.0.0",
-    "@nuxtjs/eslint-module": "^3.0.2",
-    "@nuxtjs/pwa": "^3.3.5",
-    "eslint": "^8.4.1",
-    "eslint-plugin-nuxt": "^3.1.0",
-    "eslint-plugin-vue": "^8.2.0",
-    "sass": "^1.49.0",
-    "sass-loader": "^10.2.1"
-  }
-}
