--- starters/nuxt-2/pages/404.vue
@@ -1,9 +1,0 @@
-<template>
-  <div>404</div>
-</template>
-
-<script>
-export default {
-  name: 'ErrorPage'
-};
-</script>
