+++ starters/nuxt/.env.example
@@ -1,10 +1,9 @@
 # Nacelle
-NACELLE_SPACE_ID=<your-nacelle-space-id>
-NACELLE_GRAPHQL_TOKEN=<your-nacelle-public-token>
-NACELLE_ENDPOINT=<your-nacelle-storefront-graphql-endpoint>
-NACELLE_LOCALE=<active-locale>
+NACELLE_STOREFRONT_ENDPOINT="<YOUR_NACELLE_STOREFRONT_ENDPOINT>"
+NACELLE_STOREFRONT_TOKEN="<YOUR_NACELLE_PUBLIC_STOREFRONT_TOKEN>"
+NACELLE_STOREFRONT_LOCALE="<YOUR_NACELLE_STOREFRONT_LOCALE>"
 
 # Shopify Checkout (for details, see: https://www.npmjs.com/package/@nacelle/shopify-checkout)
-MYSHOPIFY_DOMAIN=<your-shop-id>
-SHOPIFY_STOREFRONT_CHECKOUT_TOKEN=<your-storefront-api-token>
-SHOPIFY_STOREFRONT_API_VERSION=<storefront-api-version-of-choice>
+MYSHOPIFY_DOMAIN="<YOUR_MYSHOPIFY_DOMAIN>"
+SHOPIFY_STOREFRONT_CHECKOUT_TOKEN="<YOUR_SHOPIFY_STOREFRONT_CHECKOUT_TOKEN>"
+SHOPIFY_STOREFRONT_API_VERSION="<YOUR_SHOPIFY_STOREFRONT_API_VERSION>"
