+++ starters/nuxt/.eslintrc.js
@@ -11,5 +11,7 @@ module.exports = {
   extends: ['@nuxtjs', 'plugin:nuxt/recommended', 'prettier'],
   plugins: [],
   // add your custom rules here
-  rules: {}
+  rules: {
+    'no-console': 'off'
+  }
 };
