+++ starters/nuxt/components/ProductCard.vue
