+++ starters/nuxt/jsconfig.json
