+++ starters/nuxt/pages/404.vue
@@ -1,3 +1,9 @@
 <template>
   <div>404</div>
 </template>
+
+<script>
+export default {
+  name: 'ErrorPage'
+};
+</script>
