+++ starters/nuxt/pages/index.vue
@@ -14,6 +14,12 @@
   </div>
 </template>
 
+<script>
+export default {
+  name: 'HomePage'
+};
+</script>
+
 <style lang="scss" scoped>
 .home {
   align-items: center;
