+++ starters/nuxt/plugins/nacelle-sdk.js
