+++ starters/nuxt/plugins/nuxt-client-init.client.js
