+++ starters/nuxt/store/index.js
