+++ starters/nuxt/utils/buildRoutes.js
