+++ starters/nuxt/utils/getSelectedVariant.js
